requirements.txt
python
Pytorch 
